sample code/RSD_Studio_Modular_Full/src/requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
imageio-ffmpeg==0.4.9
opencv-python==4.10.0.84
pillow==10.4.0

Pillow
imageio
imageio-ffmpeg
av
